More changes to Rythmos and Thyra in support of forward sensitvities for Rythmos:
Changes to Rythmos:
-------------------

* Changed from array of RCP<VetorBase> objects to array of RCP<const VectorBase>
objects in Rythmos::InterpolationBufferBase functions setPoints(...) and
getPoints(...) and in all subclasses.  Despite having to change a lot of type
declarations, this refactoring was easier than one might think since all of
the existing code assumed that these values should not change and made copies
as needed.  Also, I never had to open a debugger since the compiler showed
all of the places that needed to be updated.

* Expanded on verbose output and status output from the
Rythmos::TimeStepNonlinearSolver::solve(...) function.

* Updated Rythmos::TimeStepNonlinearSolver to accept all if it's "magic"
parameters from the parameter list.  This means that
Rythmos::TimeStepNonlinearSolver::defaultTol(...) function no longer exists
and therefore all of the example/test programs that called this function had
to be updated to set the default tolerance through the parameter list.  This
is the price we pay for progress.

* Added missing newlines to output of Rythmos::StepStatus

* Added a test case for the increased verbosity level.  This does not really
test anything, it just sets the parameters.  Testing output is not so
easy to do unless you compare to a "golden standard" which has it's problems.

Changes to Thyra:
-----------------

* Added nommeber helper functions Thyra::nonconstProductVectorBase(...) and
productVectorBase(...) to do dynamic casting from VB interface to PVP
interface and put these in the header with the PVB interface class.

* Changed the behavior of Thyra::operator<<(...) for Thyra::SolveStatus to allow
multi-line string messages.  I used this to return more information from the
Rythmos::TimeStepNonlinearSolver::solve(...) function.

* Derived Thyra::NonlinearSolverBase from Teuchos::ParameterListAcceptor so that
now all nonlinear solvers must accept a parmeter sublist to control their
options.

* Added multi-vector versions of V_StVpV(...) and Vp_V(...) to
Thyra_MultiVectorStdOps[Decl].hpp.  I used this in the forward sensitivity
model evaluator to compute the sensitivity residual.

* Removed the class Thyra::DefaultVectorMultiVector since it was not being
used anymore and since the new Thyra::DefaultMultiVectorProductVector does
what it could do and more.

* Moved Thyra_DefaultColumnwiseMultiVector[Decl].hpp from adapter support to
client support since it is a true ANA class and is not dependent on any type
of computing environment.  Doxygen should have picked this up but it did not
seem to.

* Added the classes Thyra::DefaultMultiVectorProductVector,
Thyra::DefaultMultiVectorProductVectorSpace, and
Thyra::DefaultMultiVectorLinearOpWithSolve which implement a multi-RHS system
which are really multi-vectors but look like a single drawn out vector.  These
classes need a lot more testing but they are sufficient for what is needed by
the Rythmos forward sensitivity software.

Other Changes:
--------------

* I fixed a bug in EpetraExt::DiagonalTransientModel where the parameters where
not being handled currectly.