SUMMARY:  Added DxDp test with BDF stepper and improved error message in test. BUGS:     None submitted COMMENTS: See below
1)   Did you add any source files? No

2)   Did you update or change the build system? No

3)   Did you rerun the test suite on at least the packages that you
     changed? Yes ... There are 33 passing tests two failing tests.
     The test rythmos_basicExample_03.txt is failing by a small margin
     (see below) and the test rythmos_diagonalTransient_09.txt is failing
     due to a problem with the BDF stepper used in the context of
     the forward sensitivity stepper.

EXTENDED COMMENTS:

Here I modifed diagonalTransientMain.cpp to accept the argument
--max-restate-error to allow for some error in the recomputation of the state.
This allowed me to put in a real test for DxDp using the BDF stepper that
passes even through the recomputation of the state does not exactly match the
initial computation.  I changed the test DIAGOANL8 to be a passing test on
DxDp with the BDF stepper and added a new test DIAGONAL9 (also called
FAILING1) that shows the error in the state recomputation.  If I had added
this test in the first place, I would have seen that I had forgotten to add
the 'FD Calc' sublist and I would have not wasted so much time tracking this
non-bug down.

The test DIAGONAL9 (or FAILING or FAILING1) now shows:

    Checking that x(p,finalTime) computed as part of x_bar above is the same ...

      Testing relative error between vectors x_final and x_in_x_bar_final:
        ||x_final|| = 9.99910021336769
        ||x_in_x_bar_final|| = 9.99910021468465
        x_final = Thyra::DefaultSpmdVector<double>{spmdSpace=Thyra::DefaultSpmdVectorSpace<double>{globalDim=1,localSubDim=1,localOffset=0,comm=Teuchos::SerialComm<int>}}
          0:9.99910021336769
        x_in_x_bar_final = Thyra::DefaultSpmdVector<double>{spmdSpace=Thyra::DefaultSpmdVectorSpace<double>{globalDim=1,localSubDim=1,localOffset=0,comm=Teuchos::SerialComm<int>}}
          0:9.99910021468465
        x_final - x_in_x_bar_final = Thyra::DefaultSpmdVector<double>{spmdSpace=Thyra::DefaultSpmdVectorSpace<double>{globalDim=1,localSubDim=1,localOffset=0,comm=Teuchos::SerialComm<i
    nt>}}
          0:-1.31696609173559e-09
        Check: rel_err(x_final,x_in_x_bar_final) = 1.31708460107391e-10 <= maxRestateError = 0 : failed

This should be a binary match but is not.

I also improveed the error message in the failing basicExample.exe test which now shows:

      Leaving Rythmos::ImplicitBDFStepper<double>::takeStep(0.000313386694191412,FIXED_STEP) ...

    p=0: *** Caught standard exception of type 'std::logic_error' :

      /home/rabartl/PROJECTS/Trilinos.base/Trilinos/packages/rythmos/example/basicExample/cxx_main.cpp:451:

      Throw test that evaluated to true: normLETDiff > eps

      Error, normLETDiff = 3.86132e-11 > eps = 4.44089e-12!

This error message is much better and it shows that the test is only failing
by a very small margin.