SUMMARY:  Replaced illegal (...) with {...} BUGS:     3384 COMMENTS: See below
1)   Did you add any source files? No

2)   Did you update or change the build system? No

3)   Did you rerun the test suite on at least the packages that you
     changed? Yes

EXTENDED COMMENTS:

* Epetra_CrsMatrix.cpp: I removed illegal (...) around a macro that expands to
braces and replaced it with braces.  This is not legal ISO C++ as shown by the
compiler message:

  Epetra_CrsMatrix.cpp:817: error: ISO C++ forbids braced-groups within expressions

However, the epetra test RowMatrixTrasposer_test.exe fails on my machine and I
submitted bug 3384 to Epetra.  The epetraext test Transpose_test.exe is also
failing.  Mike Heroux said that he has also seen these tests failing but told me
to not worry about this and to just check in this change.