SUMMARY:  Put in --disable-teuchos-blasfloat to optionally remove all code that           depends on the float implementation of the BLAS
BUGS:     2886

CHECKLIST:

1) Did you add any source files? No

2) Did you update or change the build system? Yes
2.a) Did you remember to do a bootstrap and do a 'cvs add' on the
     Makefile.am files and the Makefile.in files? Yes
2.b) [Optional but recommended] Did you blow away the libraries and
     executables and then rebuild the libraries? Yes

3) Did you rerun the test suite on at least the packages that you
   changed? [???]

COMMENTS:

Here I have added a configure option --disable-teuchos-blasfloat to remove all
of the code and tests that depend on the float implementation of the BLAS.
The float BLAS and LAPACK tests in Teuchos have been failing for months on my
machine (see bug 2886 filed back on 2006-11-10).  The BLAS/float adapters and
tests are enabled by default so the behavior of Trilinos is unchanged unless
you add --disable-teuchos-blasfloat at configure time.  I also had to ifdef
out several tests in thyra when this flag is set since the float tests are
failing and so I had put in
TAC_ARG_INFORM_SUBPACKAGE_USED(teuchos,blasfloat,...) macro to catch this in
thyra.  Now the BLAS and LAPACK tests pass (which does not mean much, see
below)..

Note, however, that the BLAS and LAPACK tests where written to do a simple
comparison of the float and double versions that I had to effectively disarm
to get this to compile.  These tests are very simplistic and do not really
test the BLAS and LAPACK wrappers (other than to make sure that you do not
segfault or something).  It is really up to the clients of these wrappers to
have their own tests.  This is an SQA issue that Trilinos needs to work out
and have a policy for.