SUMMARY:  Made Epetra error output go to *Teuchos::VerboseObject::getDefaultOStream()           instead of cerr when teuchos is enabled.
BUGS:     3362

CHECKLIST:

1) Did you add any source files? No

2) Did you update or change the build system? Yes
2.a) Did you remember to do a bootstrap and do a 'cvs add' on the
     Makefile.am files and the Makefile.in files? Yes
2.b) [Optional but recommended] Did you blow away the libraries and
     executables and then rebuild the libraries? Yes

3) Did you rerun the test suite on at least the packages that you
   changed? [???]

COMMENTS:

The motivation for this commit was to get Epetra to report errors to
*Teuchos::VerboseObject::getDefaultOStream() instead of std::cerr when Teuchos
is enabled (see bug 3362 for details).  This involved adding a new static
function Epetra_Object::GetTracebackStream() to return either std::cerr or
*Teuchos::VerboseObject::getDefaultOStream() depending on an ifdef if teuchos
is defined or not.  I updated Epetra_Object::ReportError(...) and the macros
EPETRA_CHK_ERR(...), EPETRA_CHK_PTR(...), and EPETRA_CHK_REF(...) to print to
Epetra_Object::GetTracebackStream() instead of std::cerr directly.

In order to do this, I had to add a "can use" depenence of Epetra on Teuchos.
Conceptually, this just involved making a few simple changes to
epetra/configure.ac, epetra/Makefile.am and epetra/Makefile.export.epetra.in.
In reality, this was a lot of work because none of the Epetra tests or
examples used the export makefile Makefile.export.epetra.  Therefore, I had to
update about 20+ makefiles to use the export makefile correctly.  I also had
to chang makefiles in tpetra and newpackage that was not using the export
makefile system.  I had no idea this would propogate so far but there is now
way that I could have known that this would have taken so much work.

Given the amount of work this involved and the low priority of this feature,
this may not have been the best time to do this refactoring.  However, this
now opens up the possibilities of using Teuchos utilities in Epetra for things
line error handling (with exceptions) and other sorts of things.