Add two alternatives to -DRAD_AUTO_AD_Const: compilation with -DRAD_REINIT=1 or -DRAD_REINIT=2 causes use of new (distinct) schemes for retaining values after storage is reclaimed when a new computation begins after Gradcomp or Weighted_Gradcomp.  RAD_REINIT=1 takes more memory and in general requires one to call aval_reset() before starting a new computation.  (Most of the time this is not necessary, but it is needed when certain expressions appear at the start of the computation.)  RAD_REINIT=2 involves an extra test at each reference to and ADvar value (and involves more memory than the older schemes, albeit less than RAD_REINIT=1).  On a simple bilinear form evaluation (vector times known matrix times vector), for small computations that probably stay in cache, the new schemes take perhaps 2 or 2.5 time as long as the older ones; on larger computations, the relative time difference is much smaller.  The relative times for RAD_REINIT=1 versus RAD_REINIT=2 depend on the computation.
Unless RAD_DISALLOW_WANTDERIV is #defined, the new schemes introduce
a new type, ADvar_nd, for ADvar values whose partials are not of
interest.  (When RAD_DISALLOW_WANTDERIV is #defined, ADvar_nd is the
same as ADvar.) No derivative propagations happen for pieces of the
computation that only involve ADvar_nd values and constants.  If x is
an ADvar or an ADvar_nd, invoking x.Wantderiv(1) indicates that
partials w.r.t.  x are of interest, and invoking x.Wantderiv(0)
indicates that they are not of interest; x.Wantderiv() returns the
current setting (0 or 1).