SUMMARY:  Updated final documentation text printed at the end of           configure.
BUGS:     None filed

CHECKLIST:

1) Did you add any source files? No

2) Did you update or change the build system? Yes, but only
   documentation

3) Did you rerun the test suite on at least the packages that you
   changed? No, no need to for this type of change

COMMENTS:

I spotted a typo in the last bit of documentation printed at
the and of a top-level configure.  I then noticed links to
pages that no longer exist on software.sandia.gov.  I fixed
the typos and fixed the links.  The one link I could not fix
was to a page configuration_manual.html that does not seem
to exist on trilinos.sanida.gov.  I just removed the text
relating to that.

We need to merge this change back into the 8.0 release branch.