SUMMARY:Changed the NOX::Solver::Generic class for resets and removed the NOX::SolverManager handle layer.
     BUGS: 3551

     CHECKLIST:

     1)   Did you add any source files?  yes
     1.a) Did you do a 'cvs add'?  yes
     1.b) Did you add them to a Makefile.am file in some way?  yes
     1.c) Did you do a new bootstrap?  yes

     2)   Did you update or change the build system?  no

     3) Did you rerun the test suite on at least the packages that you
        changed?  yes

     COMMENTS:

1. Changed the NOX::Solver::Generic class so that there are only two reset methods that now accept a NOX::Abstract::Vector instead of a NOX::Abstract::Group.  Previously, users could really screw their solver by replacing one instance of a group in the solver with another.  Additionally, this created the awkward requirement of having to keep a Group around whenever the solver was reset with a new initial guess.  This change fixes these issues.

2. Removed the handle layer NOX::Solver::Manager from NOX.  We now use a factory to build the solvers.  Previously you would build a nox solver using:

NOX::Solver::Manager solver(group, tests, params);

This call is now handled by a factory that returns a RCP to a NOX::Solver::Generic.  So in application codes, replace the above code with:

Teuchos::RCP<NOX::Solver::Generic> solver = NOX::Solver::buildSolver(group, tests, params);

The factory is an object that can be built and called directly, but notice in the above example, that there now exists a function called buildSolver() in the NOX::Solver namespace so you don't have to deal with the factory object directly.

All tests and examples have been updated and pass on both the 8.0 and development branches except for petsc...

Russell - can you update and test the petsc directory???



(checklist version 2007-06-20)